updated the  folder structure
