+++ src/app/layout.js
@@ -1,9 +1,5 @@
-import { Suspense } from "react";
-import Header from "../components/Header";
-import SideBar from "../components/Sidebar";
-
 import "./globals.css";
-import Spinner from "../components/Spinner";
+
 
 export const metadata = {
   title: "Circle Soft",
@@ -12,17 +8,7 @@ export const metadata = {
 export default function RootLayout({ children }) {
   return (
     <html lang="en">
-      <body className="relative bg-[#f8f8f8] w-full overflow-hidden h-dvh">
-        <Header />
-        <div className="flex justify-normal items-start w-full !h-screen relative top-16">
-          <div className="h-full !max-w-[15%] !w-full !overflow-hidden">
-            <SideBar />
-          </div>
-          <div className="h-full max-w-[85%] w-full overflow-y-scroll overflow-hidden p-10 pb-24">
-            <Suspense fallback={<Spinner />}>{children}</Suspense>
-          </div>
-        </div>
-      </body>
+      <body className="">{children}</body>
     </html>
   );
 }
